+++ .gitignore
@@ -1,9 +1,9 @@
 # ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
 !/.gitattributes
-!/.projen/tasks.json
-!/.projen/deps.json
-!/.projen/files.json
-!/.github/workflows/pull-request-lint.yml
+!/.projen\tasks.json
+!/.projen\deps.json
+!/.projen\files.json
+!/.github\workflows\pull-request-lint.yml
 !/package.json
 !/LICENSE
 !/.npmignore
@@ -34,13 +34,13 @@ jspm_packages/
 /test-reports/
 junit.xml
 /coverage/
-!/.github/workflows/build.yml
+!/.github\workflows\build.yml
 /dist/changelog.md
 /dist/version.txt
-!/.github/workflows/release.yml
+!/.github\workflows\release.yml
 !/.mergify.yml
-!/.github/workflows/upgrade-master.yml
-!/.github/pull_request_template.md
+!/.github\workflows\upgrade-master.yml
+!/.github\pull_request_template.md
 !/test/
 !/tsconfig.dev.json
 !/src/
